Bug Fix: migrations fail when using in memory database
Now that the live system is using an in-memory database with a file
backup we cannot run migrations on the (in memory) database before
starting the app (because the in-memory database will be destroyed
before starting the app).

This commit fixes this issue by running migrations on the backup
database before loading it into memory. This also fixes a migration
that assumed the live database is a file and connected to the database
that the application is aware of (the in-memory database).